Escms.org
ESCMS drives the coordination of processes and activities across areas of business, sales, purchase and stock management with the help of information technology. ESCMS use to manage the demand and supply of products timely according to the user requirement. It is an automated system use to place online order, delivery and manage stock report.